Reorder files and add enhancements to mobile platforms
