+++ lib/widgets/device_grid/desktop/layout_manager.dart
